+++ main/detailed.html
@@ -57,8 +57,7 @@
                   >HOME</a
                 >
               </li>
-              <a href="./index.html"></a>
-              <li><a href="#" class="nav-link px-2 text-white">예고편</a></li>
+              <a id="homeLink" href="./index.html"></a>
             </ul>
           </div>
         </div>
