+++ main/src/detailed.js
@@ -1,6 +1,8 @@
-import { drawMovieDetails } from "./movieApi.js";
+import { drawMovieDetails } from "./movieDraw.js";
+import {movieVideo} from "./movieDraw.js";
 const check = localStorage.getItem("cast");
 drawMovieDetails(check);
+movieVideo(check);
 
 // Return to HOME
 const homeBtn = document.querySelector("#homeBtn");
