+++ main/src/movieDraw.js
@@ -1,12 +1,19 @@
+import { fetchmovieList } from "./api.js";
+import { fetchmovieCast } from "./api.js";
+import { fetchMovieDetails } from "./api.js";
+import { fetchMovieVideo } from "./api.js";
+
+
+
 // movie 카드 함수
 export const drawmovie = async () => {
-  const movieList = await fetchmovieList();
-  console.log(movieList);
+    const movieList = await fetchmovieList();
+    console.log(movieList);
 
-  const cardList = document.querySelector("#main-card");
-  cardList.innerHTML = movieList
-    .map(
-      (movie) => `
+    const cardList = document.querySelector("#main-card");
+    cardList.innerHTML = movieList
+        .map(
+            (movie) => `
       <div id="col" class="col" >
       <a href = "#">
       <div id="cardId" class="card h-100" onclick="location.href='detailed.html'">
@@ -19,47 +26,47 @@ export const drawmovie = async () => {
   </div>
   </a> 
 </div>`
-    )
-    .join("");
+        )
+        .join("");
 
-  // 카드 클릭시 페이지 넘어가고 영화 아이디 저장
-  cardList.addEventListener("click", clickCard);
+    // 카드 클릭시 페이지 넘어가고 영화 아이디 저장
+    cardList.addEventListener("click", clickCard);
 
-  function clickCard(event) {
-    if (event.traget === cardList) return;
+    function clickCard(event) {
+        if (event.traget === cardList) return;
 
-    if (event.target.matches(".image")) {
-      const id = event.target.id;
-      storegeAdd(id);
+        if (event.target.matches(".image")) {
+            const id = event.target.id;
+            storegeAdd(id);
+        }
     }
-  }
 };
 
 //로컬스토리지 저장 함수
 function storegeAdd(keyword) {
-  const data = localStorage.setItem("cast", keyword);
-  return data;
+    const data = localStorage.setItem("cast", keyword);
+    return data;
 }
 
 // id로 영화 검색후 배우 데이터 가져오기 및 카드 추가
 export const drawMovieDetails = async (id) => {
-  const detailsList = await fetchMovieDetails(id);
-  const detailedCast = await fetchmovieCast(id);
+    const detailsList = await fetchMovieDetails(id);
+    const detailedCast = await fetchmovieCast(id);
 
-  const cardDetailList = document.querySelector("#moviedetail");
-  const moviedetailCast = document.querySelector("#moviedetailCast");
+    const cardDetailList = document.querySelector("#moviedetail");
+    const moviedetailCast = document.querySelector("#moviedetailCast");
 
-  let backdrop = detailsList.backdrop_path;
-  let title = detailsList.title;
-  let overview = detailsList.overview;
-  let average = detailsList.vote_average;
-  let genres = detailsList.genres.map((e) => e.name).join(", ");
-  let release = detailsList.release_date;
-  let character = detailedCast.map((e) => e.character).join("역, ");
-  let caster = detailedCast.map((e) => e.original_name).join(", ");
+    let backdrop = detailsList.backdrop_path;
+    let title = detailsList.title;
+    let overview = detailsList.overview;
+    let average = detailsList.vote_average;
+    let genres = detailsList.genres.map((e) => e.name).join(", ");
+    let release = detailsList.release_date;
+    let character = detailedCast.map((e) => e.character).join("역, ");
+    let caster = detailedCast.map((e) => e.original_name).join(", ");
 
 
-  let tempHtml = `<div class="main" img
+    let tempHtml = `<div class="main" img
   style="
   background: linear-gradient(to bottom, rgba(0, 0, 0, 0) 85%, rgba(0, 0, 0, 0.7) 95%, rgba(0, 0, 0, 1) 100%), url(https://image.tmdb.org/t/p/w500${backdrop}); background-size: cover; background-position: center; background-repeat: no-repeat;">
   <div class="p-5 mb-4 bg-body-tertiary rounded-3">
@@ -73,9 +80,9 @@ export const drawMovieDetails = async (id) => {
       <button id="rewiewbtn" type="button" class="btn btn-outline-light">리뷰</button>
   </div>
 </div>`;
-  cardDetailList.insertAdjacentHTML("afterbegin", tempHtml);
+    cardDetailList.insertAdjacentHTML("afterbegin", tempHtml);
 
-  let tempCastHtml = `<div
+    let tempCastHtml = `<div
 style="border: whitesmoke 0px solid; border-radius: 15px; background-color: rgb(0, 0, 0); margin: 5px 200px 5px 280px; padding: 10px 50px 10px 50px;">
 
 <div class="moviesText">
@@ -139,72 +146,20 @@ style="border: whitesmoke 0px solid; border-radius: 15px; background-color: rgb(
     </div>
 </div>
 </div>`;
-  moviedetailCast.insertAdjacentHTML("beforeend", tempCastHtml);
+    moviedetailCast.insertAdjacentHTML("beforeend", tempCastHtml);
 };
 
-// API MOVIE 데이터 함수
-async function fetchmovieList() {
-  const options = {
-    method: "GET",
-    headers: {
-      accept: "application/json",
-      Authorization:
-        "Bearer eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiJhMzBjZjNhMjBlZGNhZjMxMmIwMjZhZjM1NzhiMTAyOCIsInN1YiI6IjY1OTRmNDQ5NTkwN2RlNDU5OTYzYmZkNSIsInNjb3BlcyI6WyJhcGlfcmVhZCJdLCJ2ZXJzaW9uIjoxfQ.jIu9EyJ2GTlw8ENNNk9IuC76BKQ_Ii0J4QAWX_-Jo00",
-    },
-  };
-
-  try {
-    const response = await fetch(
-      "https://api.themoviedb.org/3/discover/movie?include_adult=false&include_video=false&language=ko-KO&page=1&sort_by=popularity.desc",
-      options
-    );
-    const data = await response.json();
-    return data.results;
-  } catch (err) {
-    console.error(err);
-  }
-}
-
-// API CAST 데이터 함수
-async function fetchmovieCast(id) {
-  const options = {
-    method: "GET",
-    headers: {
-      accept: "application/json",
-      Authorization:
-        "Bearer eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiJhMzBjZjNhMjBlZGNhZjMxMmIwMjZhZjM1NzhiMTAyOCIsInN1YiI6IjY1OTRmNDQ5NTkwN2RlNDU5OTYzYmZkNSIsInNjb3BlcyI6WyJhcGlfcmVhZCJdLCJ2ZXJzaW9uIjoxfQ.jIu9EyJ2GTlw8ENNNk9IuC76BKQ_Ii0J4QAWX_-Jo00",
-    },
-  };
-  try {
-    const response = await fetch(
-      `https://api.themoviedb.org/3/movie/${id}/credits?language=ko-KO`,
-      options
-    );
-    const data = await response.json();
-    return data.cast;
-  } catch (err) {
-    console.error(err);
-  }
-}
-
-async function fetchMovieDetails(id) {
-  const options = {
-    method: "GET",
-    headers: {
-      accept: "application/json",
-      Authorization:
-        "Bearer eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiJhMzBjZjNhMjBlZGNhZjMxMmIwMjZhZjM1NzhiMTAyOCIsInN1YiI6IjY1OTRmNDQ5NTkwN2RlNDU5OTYzYmZkNSIsInNjb3BlcyI6WyJhcGlfcmVhZCJdLCJ2ZXJzaW9uIjoxfQ.jIu9EyJ2GTlw8ENNNk9IuC76BKQ_Ii0J4QAWX_-Jo00",
-    },
-  };
-
-  try {
-    const response = await fetch(
-      `https://api.themoviedb.org/3/movie/${id}?language=ko-KO`,
-      options
-    );
-    const data = await response.json();
-    return data;
-  } catch (err) {
-    console.error(err);
-  }
+export const movieVideo = async (id) => {
+    const movieVideo = await fetchMovieVideo(id);
+    console.log(movieVideo);
+    let key = '';
+    let check = "Trailer"
+    for (let i = 0; i < movieVideo.length; i++) {
+        if (movieVideo[i].type === check) {
+            key += movieVideo[i].key;
+        }
+    }
+    const videoLink = document.querySelector("#homeLink");
+    let tempVideo_html = `<li><a href="https://www.youtube.com/watch?v=${key}" class="nav-link px-2 text-white">예고편</a></li>`;
+    videoLink.insertAdjacentHTML("afterbegin", tempVideo_html);
 }
